+++ scrapyd/jobstorage.py
@@ -15,6 +15,21 @@ class Job:
         self.start_time = start_time if start_time else datetime.datetime.now()
         self.end_time = end_time if end_time else datetime.datetime.now()
 
+    def __eq__(self, other):
+        return (
+            self.project == other.project
+            and self.spider == other.spider
+            and self.job == other.job
+            and self.start_time == other.start_time
+            and self.end_time == other.end_time
+        )
+
+    def __repr__(self):
+        return (
+            f"Job(project={self.project}, spider={self.spider}, job={self.job}, "
+            f"start_time={self.start_time}, end_time={self.end_time})"
+        )
+
 
 @implementer(IJobStorage)
 class MemoryJobStorage:
@@ -27,13 +42,13 @@ class MemoryJobStorage:
         del self.jobs[: -self.finished_to_keep]  # keep last x finished jobs
 
     def list(self):
-        return self.jobs
+        return list(self)
 
     def __len__(self):
         return len(self.jobs)
 
     def __iter__(self):
-        yield from self.jobs
+        yield from reversed(self.jobs)
 
 
 @implementer(IJobStorage)